codigo não otimizado com long double

+++ src/broyden.h
@@ -3,15 +3,15 @@
 
 #include "utils.h"
 
-void montaJacobiana(const double *x, double **j, int n);
+void montaJacobiana(const long double *x, long double **j, int n);
 
-void avaliaF(const double *x, double *f, int n);
+void avaliaF(const long double *x, long double *f, int n);
 
 //newton otimizado
-//void newton(double *a, double *b, double *c, double *f, double *s, double *x, double tol, int max_iter, int n, rtime_t tempoJ, rtime_t tempo_resolucao);
+//void newton(long double *a, long double *b, long double *c, long double *f, long double *s, long double *x, long double tol, int max_iter, int n, rtime_t tempoJ, rtime_t tempo_resolucao);
 
-void newton(double **j, double *f, double *s, double *x, double tol, int max_iter, int n, rtime_t *tempoJ, rtime_t *tempo_resolucao);
+void newton(long double **j, long double *f, long double *s, long double *x, long double tol, int max_iter, int n, rtime_t *tempoJ, rtime_t *tempo_resolucao);
 
-double norma(const double *v, int n);
+long double norma(const long double *v, int n);
 
 #endif
